FLASK_PROJECT/todo_list_v2/requirements.txt
flask
flask-cors
bcrypt
pyjwt